Rename folders, mainContent layout
