--- application/src/mainLayout/components/MainContent.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const MainContent = () => {
-    return (
-        <div>MainContent</div>
-    )
-}
-
-
-export default MainContent;
